connection-manager: server
On outound connection we need to be able to manage responders: not only
start them but also re-start them when they are finished.  The server
knows how to do it! This patch provides `ControlChannel`, which allows
to notifiy the server about a thread that it needs to monitor (start
& re-start responders when they finish).

This patch also extends the tests, which now runs multiple rounds of
`ReqResp` protocol, which forces the server to use the restarting logic.